Risen 2: Fix point lights with shadows
Fairly straight forward world space correction in vertex + pixel shader.
The main twist was that r1 needs to be corrected in the pixel shader as
well as it is used throughout the rest of the shader to place the light.

I used c210 for the inverse view-projection matrix instead of c220, so
that I could use DoubleInverseMatrix to pass the forwards
view-projection matrix as well. It may be possible to skip this if r0.z
is the depth - I haven't checked.